Change id type from Integer to Long
JIRA:
CG-78

+++ src/main/java/org/cookbookgeeks/webkochbuch/dao/CategoryDao.java
@@ -26,7 +26,7 @@ import org.cookbookgeeks.webkochbuch.domain.Category;
  * @author Nils Sommer
  *
  */
-public class CategoryDao extends GenericDao<Category, Integer> {
+public class CategoryDao extends GenericDao<Category, Long> {
 	
 	/**
 	 * Standard constructor.
